refactor : useRefLocate로 custom hook으로 떼넴
